Hold data for all relations in a single InMemoryLayer.
This is a small step towards turning the storage format into an LSM tree.
In an LSM tree, the MemTable - or InMemoryLayer in our terms - holds data
for the whole key range.  Checkpoint still writes it out to multiple
per-segment layers, however, so this commit the on-disk format yet.

Move code to handle rel size updates from InMemoryLayer to the caller.
When you modify a block of a relish or segment for the first time,
the InMemoryLayer doesn't know its old size. The caller needs to find it
out and tell it. This fixes the residual problem with zero-filling "gaps"
when a relation is extended at a segment boundary (see #500). There was a
test case for that in repository.rs, but it was commented out because it
didn't work before. It is now enabled.

Merge InMemoryLayer::put_wal_record and put_page_image into a single
put_page_version function.